refactor: move provider-management back into core + trim module glue
Same reasoning that moved scheduling back to core in e41822b5 applies to
provider-management: models sit on the critical path of 100% of runs, the
"module" reached into core through 15 ../../../ imports, and resolveModel
had a single implementer so the hook was indirection for nothing.
Disabling the module also dropped features.models/providerKeys, breaking
the BYO-key UI for every OSS self-hoster by default — the opposite of what
a core toolbox should do. Only webhooks remains a module (clean
onRunStatusChange listener, zero reach-back, isolated BullMQ worker).

- Move routes, services, openapi paths, schema, migration out of
  apps/api/src/modules/provider-management/ into core (routes/,
  services/, openapi/paths/, packages/db/src/schema/provider-keys.ts,
  packages/db/drizzle/0004_provider_management_into_core.sql with
  CREATE TABLE IF NOT EXISTS for safe rollout).
- Remove resolveModel from ModuleHooks in @appstrate/core/module;
  env-builder.ts now calls services/org-models.ts#resolveModel
  directly. Delete resolveSystemModel (dead after inlining).
- Inline lib/modules/hooks.ts into its 5 callsites
  (boot, run-pipeline, runs, scheduler, organizations) as direct
  callHook/emitEvent calls, then delete the file — consistent with
  env-builder which already called callHook directly, and the
  dedicated test file already deleted in f850da4b.
- Drop features.models and features.providerKeys everywhere
  (buildAppConfig, AppConfig type, use-app-config defaults,
  onboarding-layout StepDef, 5 UI guards) — symmetric to ac2761cc.
- Update CLAUDE.md, modules/README.md, test helpers.

Net: +262 / -2375 across 42 files. bun run check green, bun test
1899 pass / 0 fail.
