+++ apps/api/src/lib/app-config.ts
@@ -14,8 +14,6 @@ export function buildAppConfig(): AppConfig {
   return {
     features: {
       billing: false,
-      models: false,
-      providerKeys: false,
       webhooks: false,
       googleAuth: !!(env.GOOGLE_CLIENT_ID && env.GOOGLE_CLIENT_SECRET),
       githubAuth: !!(env.GITHUB_CLIENT_ID && env.GITHUB_CLIENT_SECRET),
