+++ apps/api/src/modules/README.md
@@ -1,6 +1,6 @@
 # Built-in Modules
 
-Built-in modules extend the Appstrate platform with optional features (webhooks, provider management, …). They follow the same `AppstrateModule` contract as external modules like `@appstrate/cloud`, but live inside the API package so they can share test infrastructure and be discovered automatically.
+Built-in modules extend the Appstrate platform with optional features (currently: webhooks). They follow the same `AppstrateModule` contract as external modules like `@appstrate/cloud`, but live inside the API package so they can share test infrastructure and be discovered automatically.
 
 ## Auto-discovery
 
@@ -65,11 +65,11 @@ Everything else (`hooks`, `events`, `openApiComponentSchemas`, `openApiSchemas`,
 1. Module tables live in `apps/api/src/modules/<id>/schema.ts`. They have their own Drizzle migration tree and a dedicated tracking table `__drizzle_migrations_<id>` (hyphens replaced with underscores).
 2. **Backward FKs (module → core)** use Drizzle `.references()` inline in the module schema — for example `orgId: uuid("org_id").notNull().references(() => organizations.id, { onDelete: "cascade" })`. Core tables always exist before any module migration runs at boot, so this is safe.
 3. **Forward FKs (core → module)** cannot be expressed via Drizzle without leaking the module schema into core. When you need one, add it via raw SQL inside the module's own migration. Core stays agnostic of module tables.
-4. Core never imports from `apps/api/src/modules/`. If core code needs data from a module, use a hook (`resolveModel`, `beforeRun`, `afterRun`) — never a direct import.
+4. Core never imports from `apps/api/src/modules/`. If core code needs data from a module, use a hook (`beforeRun`, `afterRun`) — never a direct import.
 
 ## Permissions
 
-RBAC is a platform capability, not a module concern. Core's `apps/api/src/lib/permissions.ts` is the single typed source of truth for the `Permission` taxonomy — it already lists every resource the built-in modules use (`webhooks`, `models`, `provider-keys`), together with the role-to-permission matrix and the API key allowlist. Module manifests do not declare permissions or scopes.
+RBAC is a platform capability, not a module concern. Core's `apps/api/src/lib/permissions.ts` is the single typed source of truth for the `Permission` taxonomy — it already lists every resource the built-in modules use (`webhooks`), together with the role-to-permission matrix and the API key allowlist. Module manifests do not declare permissions or scopes.
 
 Inside module routes, protect handlers with the same typed helper core uses: `requirePermission("webhooks", "write")` (from `apps/api/src/middleware/require-permission.ts`). TypeScript will reject any resource/action pair that is not in `ResourceActions` — if you need a new resource, add it to `permissions.ts` in the same PR that adds the module. Treat this as part of the platform contract: core ships the vocabulary, modules implement the behavior.
 
@@ -77,7 +77,7 @@ The trade-off is that disabling a module leaves a handful of inert permission st
 
 ## Hooks and events
 
-- **Hooks** (`callHook`, first-match-wins): `beforeRun`, `afterRun`, `beforeSignup`, `resolveModel`. The first module that provides a hook is called, subsequent modules are skipped. `beforeRun` gates a run, `afterRun` returns a metadata patch persisted on the final run record, `beforeSignup` gates signup, `resolveModel` maps an agent to an LLM config.
+- **Hooks** (`callHook`, first-match-wins): `beforeRun`, `afterRun`, `beforeSignup`. The first module that provides a hook is called, subsequent modules are skipped. `beforeRun` gates a run, `afterRun` returns a metadata patch persisted on the final run record, `beforeSignup` gates signup.
 - **Events** (`emitEvent`, broadcast-to-all): `onRunStatusChange`, `onOrgCreate`, `onOrgDelete`. Handlers run for side effects only; errors in one handler are isolated and do not block others.
 
 Names are defined in `packages/core/src/module.ts` (`ModuleHooks`, `ModuleEvents`). To add a new hook or event, update that file first so both platform and modules see the same contract.
