+++ apps/api/src/openapi/paths/provider-keys.ts
