+++ apps/api/src/openapi/zod-schema-registry.ts
@@ -8,9 +8,9 @@
  * JSON Schemas against the hand-written OpenAPI requestBody schemas.
  *
  * Core schemas are defined statically here. Module-owned schemas
- * (schedules, webhooks, models, provider-keys) are contributed
- * dynamically via `openApiSchemas()` — they only appear when the
- * module is loaded. Call `buildZodSchemaRegistry()` after module init.
+ * (webhooks) are contributed dynamically via `openApiSchemas()` — they
+ * only appear when the module is loaded. Call `buildZodSchemaRegistry()`
+ * after module init.
  */
 
 import { z } from "zod";
@@ -94,6 +94,21 @@ import {
   updatePackageSchema,
   appProviderCredentialsSchema,
 } from "../routes/applications.ts";
+
+// --- Models schemas (routes/models.ts) ---
+import {
+  createModelSchema,
+  updateModelSchema,
+  setDefaultSchema as modelsSetDefaultSchema,
+  testInlineSchema as modelsTestInlineSchema,
+} from "../routes/models.ts";
+
+// --- Provider key schemas (routes/provider-keys.ts) ---
+import {
+  createSchema as createProviderKeySchema,
+  updateSchema as updateProviderKeySchema,
+  testInlineSchema as providerKeysTestInlineSchema,
+} from "../routes/provider-keys.ts";
 
 // ---------------------------------------------------------------------------
 // Registry type and entries
@@ -405,6 +420,52 @@ const coreSchemas: ZodSchemaEntry[] = [
     path: "/internal/connections/report-auth-failure",
     jsonSchema: toJsonSchema(reportAuthFailureSchema),
     description: "Report auth failure from sidecar",
+  },
+
+  // ─── Models ─────────────────────────────────────────────────────────────
+  {
+    method: "POST",
+    path: "/api/models",
+    jsonSchema: toJsonSchema(createModelSchema),
+    description: "Create model",
+  },
+  {
+    method: "PUT",
+    path: "/api/models/{id}",
+    jsonSchema: toJsonSchema(updateModelSchema),
+    description: "Update model",
+  },
+  {
+    method: "PUT",
+    path: "/api/models/default",
+    jsonSchema: toJsonSchema(modelsSetDefaultSchema),
+    description: "Set default model",
+  },
+  {
+    method: "POST",
+    path: "/api/models/test",
+    jsonSchema: toJsonSchema(modelsTestInlineSchema),
+    description: "Test model config inline",
+  },
+
+  // ─── Provider Keys ──────────────────────────────────────────────────────
+  {
+    method: "POST",
+    path: "/api/provider-keys",
+    jsonSchema: toJsonSchema(createProviderKeySchema),
+    description: "Create provider key",
+  },
+  {
+    method: "PUT",
+    path: "/api/provider-keys/{id}",
+    jsonSchema: toJsonSchema(updateProviderKeySchema),
+    description: "Update provider key",
+  },
+  {
+    method: "POST",
+    path: "/api/provider-keys/test",
+    jsonSchema: toJsonSchema(providerKeysTestInlineSchema),
+    description: "Test provider key inline",
   },
 ];
