+++ apps/api/test/helpers/seed.ts
@@ -27,9 +27,10 @@ import {
   packageVersions,
   applicationProviderCredentials,
   packageSchedules,
+  orgProviderKeys,
+  orgModels,
 } from "@appstrate/db/schema";
 import { webhooks } from "../../src/modules/webhooks/schema.ts";
-import { orgProviderKeys, orgModels } from "../../src/modules/provider-management/schema.ts";
 import type { InferInsertModel, InferSelectModel } from "drizzle-orm";
 
 // ─── Packages / Agents ───────────────────────────────────
