+++ apps/web/src/pages/onboarding/done-step.tsx
@@ -73,24 +73,22 @@ export function OnboardingDoneStep() {
           </div>
         )}
 
-        {features.models && (
-          <div className="border-border bg-card rounded-lg border p-4">
-            <div className="flex items-center gap-3">
-              <CheckCircle2
-                size={20}
-                className={
-                  defaultModel ? "shrink-0 text-green-500" : "text-muted-foreground shrink-0"
-                }
-              />
-              <div className="flex-1">
-                <h3 className="text-sm font-semibold">{t("onboarding.summaryModel")}</h3>
-                <span className="text-muted-foreground text-sm">
-                  {defaultModel ? defaultModel.label : t("onboarding.summaryModelNone")}
-                </span>
-              </div>
+        <div className="border-border bg-card rounded-lg border p-4">
+          <div className="flex items-center gap-3">
+            <CheckCircle2
+              size={20}
+              className={
+                defaultModel ? "shrink-0 text-green-500" : "text-muted-foreground shrink-0"
+              }
+            />
+            <div className="flex-1">
+              <h3 className="text-sm font-semibold">{t("onboarding.summaryModel")}</h3>
+              <span className="text-muted-foreground text-sm">
+                {defaultModel ? defaultModel.label : t("onboarding.summaryModelNone")}
+              </span>
             </div>
           </div>
-        )}
+        </div>
 
         <div className="border-border bg-card rounded-lg border p-4">
           <div className="flex items-center gap-3">
