+++ packages/db/drizzle/meta/_journal.json
@@ -29,6 +29,13 @@
       "when": 1775656830201,
       "tag": "0003_overrated_ozymandias",
       "breakpoints": true
+    },
+    {
+      "idx": 4,
+      "version": "7",
+      "when": 1776067200000,
+      "tag": "0004_provider_management_into_core",
+      "breakpoints": true
     }
   ]
 }
